requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
biopython=1.79=pypi_0
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py38hecd8cb5_2
libcxx=12.0.0=h2f01273_0
libffi=3.2.1=1
ncurses=6.3=hca72f7f_2
numpy=1.22.1=pypi_0
openssl=1.1.1m=hca72f7f_0
pip=21.2.4=py38hecd8cb5_0
python=3.8.2=hc70fcce_0
readline=8.1.2=hca72f7f_1
setuptools=58.0.4=py38hecd8cb5_0
sqlite=3.37.0=h707629a_0
tk=8.6.11=h7bc2e8c_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h4dc903c_4
